# Frontend/public/api/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "chatbox-server",
        "description": "chatbox 后端 API 文档",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:5100",
            "description": "开发环境"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": ["用户认证"],
                "summary": "用户登录",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserPass"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Token"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "身份认证失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginFailed"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": ["用户认证"],
                "summary": "用户注册",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserPass"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "注册成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Token"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "注册失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegisterFailed"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/refresh_token": {
            "post": {
                "tags": ["用户认证"],
                "summary": "刷新JWT",
                "operationId": "refreshToken",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SingleToken"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "刷新成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Token"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "刷新失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RefreshFailed"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/getProfile": {
            "get": {
                "tags": ["个人信息"],
                "summary": "获取个人信息",
                "operationId": "getProfile",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/TokenExpired"
                    }
                }
            }
        },
        "/user/groups/list": {
            "get": {
                "tags": ["群聊管理"],
                "summary": "获取群组列表",
                "operationId": "getGroups",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GroupsList"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/TokenExpired"
                    }
                }
            }
        },
        "/user/groups/join": {
            "post": {
                "tags": ["群聊管理"],
                "summary": "加入群组",
                "operationId": "joinGroup",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "group_id": 10231
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/TokenExpired"
                    }
                }
            }
        },
        "/user/groups/create": {
            "post": {
                "tags": ["群聊管理"],
                "summary": "创建群组",
                "operationId": "createGroup",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "group_name": "关于YJX如何AKIOI的讨论群"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/TokenExpired"
                    }
                }
            }
        },
        "/user/groups/leave": {
            "delete": {
                "tags": ["群聊管理"],
                "summary": "退出群组",
                "operationId": "leaveGroup",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "group_id": 114451
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/TokenExpired"
                    }
                }
            }
        },
        "/user/friends/list": {
            "get": {
                "tags": ["好友管理"],
                "summary": "获取好友列表",
                "operationId": "getFriends",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FriendsList"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/TokenExpired"
                    }
                }
            }
        },
        "/user/friends/add": {
            "post": {
                "tags": ["好友管理"],
                "summary": "添加好友",
                "operationId": "addFriend",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "friend_id": 10231
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/TokenExpired"
                    }
                }
            }
        },
        "/user/friends/delete": {
            "delete": {
                "tags": ["好友管理"],
                "summary": "删除好友",
                "operationId": "deleteFriend",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "friend_id": 114451
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/TokenExpired"
                    }
                }
            }
        }
    },
    
    "components": {
        "schemas": {
            "UserPass": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "admin"
                    },
                    "password": {
                        "type": "string",
                        "example": "123456"
                    }
                }
            },
            "Token": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                    },
                    "expires": {
                        "type": "integer",
                        "example": 3600
                    }
                }
            },
            "SingleToken": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                    }
                }
            },
            "Unauthorized": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 401
                    },
                    "message": {
                        "type": "string",
                        "example": "登录信息已过期，请重新登录"
                    }
                }
            },
            "Success": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 200
                    },
                    "message": {
                        "type": "string",
                        "example": "操作成功"
                    }
                }
            },
            "LoginFailed": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 401
                    },
                    "message": {
                        "type": "string",
                        "example": "用户名或密码错误"
                    }
                }
            },
            "RegisterFailed": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 422
                    },
                    "message": {
                        "type": "string",
                        "example": "用户名已存在"
                    }
                }
            },
            "RefreshFailed": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 422
                    },
                    "message": {
                        "type": "string",
                        "example": "Token刷新失败"
                    }
                }
            },
            "GroupsList": {
                "type": "object",
                "properties": {
                    "groups": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 10231
                                },
                                "name": {
                                    "type": "string",
                                    "example": "24级院系资料群"
                                },
                                "avatar": {
                                    "type": "string",           
                                    "example": "avatar/bGciR5cCI6IkpXsC.jpg"
                                }
                            }
                        }
                    }
                }
            },
            "FriendsList": {
                "type": "object",
                "properties": {
                    "friends": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 114514
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Iron_china"
                                },
                                "avatar": {
                                    "type": "string",           
                                    "example": "avatar/WTY_AK_IOI.jpg"
                                }
                            }
                        }
                    }
                }
            },
            "Profile": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10231
                    },
                    "username": {
                        "type": "string",
                        "example": "admin"
                    },
                    "avatar": {
                        "type": "string",           
                        "example": "avatar/bGciR5cCI6IkpXsC.jpg"
                    }
                }
            }
        },
        "responses": {
            "TokenExpired": {
                "description": "token过期或不存在",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Unauthorized"
                        }
                    }
                }
            },
            "Success": {
                "description": "操作成功",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Success"
                        }
                    }
                }
            }
        }
    }
}